replace use of multi-stage docker build
instead of unsing the multi-stage docker build, which will download
the sdk artifacts once for each build, we should download the sdk
artifacts using a separate gitlab ci job and copy the cmdline-tools
into the container.

this change will allow us to reduce the build time for each container
significantly.

closes #7